[android] #5004 - Allowing MapboxAccountManager and MapView to both initialize telemetry depending on developer implementation
